'remove unused materials' function
Checks for materials with no vertex assignment and removes them from the selected objects.

This should probably be debugged extensively since it is directly removing user data, which we should be careful handling.

Currently it does warn and suggests adding a Fake User, so Blender doesn't remove the zero-user material data blocks.