re-add regressed solid mode DOF function
Not sure how this got lost. Should work now, even when in a different viewport mode, which is good.

Also, we now deselect everything before running the operator. We might be able to update this later if need be, it's not an ideal fix to the incorrec context application for the camera_to_view but it'll do