Wireframe roggle, camera category and functions, Remove Unused Material function, general cleanup (#14)
* Add viewport wireframe, add viewport camera setup

* Update vv_tools.py

* Update vv_tools.py

* Fix menus

* Boxify rig xfer button

* 'remove unused materials' function

Checks for materials with no vertex assignment and removes them from the selected objects.

This should probably be debugged extensively since it is directly removing user data, which we should be careful handling.

Currently it does warn and suggests adding a Fake User, so Blender doesn't remove the zero-user material data blocks.

* Update README.md

* re-add regressed solid mode DOF function

Not sure how this got lost. Should work now, even when in a different viewport mode, which is good.

Also, we now deselect everything before running the operator. We might be able to update this later if need be, it's not an ideal fix to the incorrec context application for the camera_to_view but it'll do

* Updated readme in general for viewport cam

* Update README.md

* Update README.md

* Update README.md

* Update README.md