Fix texture reloader (#20)
* Add VVtools_module folder, for developing modular

* start of manual module push...

* TIL refactoring

Mostly unfucked!

We have moved to a proper file structure, separating menus / operators / panels into submodules.
Is this normally hard? I feel like it should have been more straightforward. Hey ho.

* Fix the file structure, include the zipped addon for ease of release

Was previously in a _module folder for dev. This change makes it appropriate to push to main.

* small readme tweak, add version to zipped addon

* Add preference menu update checker

Uses code from https://github.com/CGCookie/blender-addon-updater

It's not particularly fanciful, but it is integrated and it does appear to work for now!

* Update README.md

* ReloadTexturesOfSelected fix